requirements.txt
pymongo~=3.11.2
numpy~=1.19.4
searchtweets~=1.7.6
dnspython

tweet-preprocessor
nltk~=3.5

scikit-learn~=0.23.2
pandas~=1.1.5
matplotlib~=3.3.3